% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{animante_deriv}
\alias{animante_deriv}
\title{View Animated Derivative Curves}
\usage{
animante_deriv(norm_data)
}
\arguments{
\item{norm_data}{dataset input, analyzed must norm_deriv as a variable;
dataset qualifying norm_data or tsar_data both fulfills this parameter}
}
\value{
plotly object of derivative curves by well
}
\description{
Function reviews data by well and output animated graph first derivatives
  Content are animated by Well IDs. Potentiall, when experiment are set wth
  concentration gradients down the plate, graph would reflect than change
  induced by such change.
}
\examples{
data("example_tsar_data")
animante_deriv(example_tsar_data)

}
\seealso{
Other TSA Plots: 
\code{\link{TSA_boxplot}()},
\code{\link{TSA_compare_plot}()},
\code{\link{TSA_wells_plot}()},
\code{\link{get_legend}()},
\code{\link{graph_tsar}()},
\code{\link{view_deriv}()}
}
\concept{TSA Plots}
