% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{download_template}
\alias{download_template}
\title{Download Well Information Template File}
\usage{
download_template()
}
\value{
None
}
\description{
Downloads Well Information Template File to the current working directory.
}
\examples{
\dontrun{
download_file()
}
}
