% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tsar_data.R
\name{example_tsar_data}
\alias{example_tsar_data}
\title{Example tsar_data file}
\format{
A data frame with the following columns:
    \describe{
      \item{Well}{Well position}
      \item{Temperature}{Temperature in degrees}
      \item{Fluorescence}{Fluorescence reading}
      \item{Normalized}{Normalized value}
      \item{norm_deriv}{Calculated first derivative}
      \item{Tm}{Tm value}
      \item{Protein}{Protein information}
      \item{Ligand}{Ligand information}
      \item{ExperimentFileName}{Experiment file name}
      \item{well_ID}{Well ID}
      \item{condition_ID}{Condition ID}
    }
}
\source{
experimentally obtained
}
\usage{
data(example_tsar_data)
}
\value{
example tsar_data in data frame
}
\description{
Dataset Description:
    This is an example dataset of the tsar_data strucutre. The data
    frame contains well ID, conditions, and experimental details.
}
\keyword{dataset}
