% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLegend.R
\name{get_legend}
\alias{get_legend}
\title{Extract ggplot2 legend}
\usage{
get_legend(input_plot)
}
\arguments{
\item{input_plot}{a ggplot2 object}
}
\value{
two ggplots, one containing the legend and another containg all else.
}
\description{
A function to enable separation of legends from plots within the TSAR package
}
