% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_tsar.R
\name{graph_tsar}
\alias{graph_tsar}
\title{Graph tsar_data}
\usage{
graph_tsar(tsar_data = data.frame())
}
\arguments{
\item{tsar_data}{tsar data outputted by norm_to_tsar or merge_tsa}
}
\value{
prompts separate app window for user interaction,
  does not return specific value; generates boxplot and compare plots
  according to user input
}
\description{
The weed_raw function allows users to interact with a screening graph
  and select curves to weed out before entering analysis. Function wraps
  together \code{\link[TSAR]{screen}} and \code{\link{remove_raw}}.
}
\examples{
# graph_tsar()

}
\seealso{
\code{\link{TSA_boxplot}}, \code{\link{tsa_compare_plot}},
  \code{\link{condition_IDs}}, \code{\link{well_IDs}}

Other TSA Plots: 
\code{\link{TSA_boxplot}()},
\code{\link{TSA_wells_plot}()},
\code{\link{get_legend}()},
\code{\link{tsa_compare_plot}()}
}
\concept{TSA Plots}
