% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_to_tsar.R
\name{merge_norm}
\alias{merge_norm}
\title{Merge and format data for sequential applications of graphing functions}
\usage{
merge_norm(
  norm_data1,
  norm_data2,
  file_name1,
  file_name2,
  exp_date1,
  exp_date2
)
}
\arguments{
\item{norm_data1}{data frame; normalized data outputted by read_tsar
following gam_analysis. Input data for experiment replication day 1}

\item{norm_data2}{data frame; normalized data outputted by read_tsar
following gam_analysis. Input data for experiment replication day 2}

\item{file_name1}{character string; input file name of experiment day 1
for data tracing purpose}

\item{file_name2}{character string; input file name of experiment day 2
for data tracing purpose}

\item{exp_date1}{character string; input date of experiment}

\item{exp_date2}{character string; input date of experiment}
}
\description{
This function merges data of experiment replicates across different dates.
  It merges and produces information variables used to group wells of same
  set up.
}
