% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{model_gam}
\alias{model_gam}
\title{modeling function}
\usage{
model_gam(Data, x, y)
}
\arguments{
\item{Data}{data frame input}

\item{x}{temperature column}

\item{y}{normalized fluorescence column}
}
\value{
gam model fitted with formula set to y ~ s(x, bs = "cs")
}
\description{
function finds fitted fluorescence values
}
