% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocess.R
\name{modelfit}
\alias{modelfit}
\title{refit and calculate derivative function
Modelfit calculates derivatives by refitting model onto data}
\usage{
modelfit(data, model)
}
\arguments{
\item{data}{data frame; the raw data set input}

\item{model}{fitted model containing fitted values}
}
\value{
data frame; with calculated derivative columns
}
\description{
refit and calculate derivative function
Modelfit calculates derivatives by refitting model onto data
}
\examples{
#fitted <- modelfit(data = gammodel, model = fitted)

}
