% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCR_data1.R
\name{qPCR_data1}
\alias{qPCR_data1}
\title{qPCR_data1 Dataset}
\format{
A data frame with the following columns:
    \describe{
      \item{Well}{Well Count, not required for user}
      \item{Well.Position}{Well Label, i.e. A01; required input}
      \item{Reading}{reading count in time series, not required for user}
      \item{Temperature}{temperature reading, required input}
      \item{Fluorescence}{fluorescence reading, required input}
    }
}
\source{
experimentally obtained
}
\usage{
data(qPCR_data1)
}
\value{
qPCR_data1 data frame
}
\description{
Dataset Description:
    This dataset contains qPCR data for the CA121 protein and common
    vitamins. It provides fluorescence measurements obtained
    using QuantStudio3.
    Dataset is experimentally obtained by author of this package.
}
\keyword{dataset}
