% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_analysis.R
\name{read_tsar}
\alias{read_tsar}
\title{read result}
\usage{
read_tsar(result, code)
}
\arguments{
\item{result}{list; input uses resulting output of gam_analysis() function}

\item{code}{integer; \code{code = 0} returns only the tm value by wells
\code{code = 1} returns data table with fitted values
\code{code = 2} returns the combination of 0 and 1}
}
\value{
output files with select dataset
}
\description{
reads previous pipeline output lists and organizes them into separate
  data frames
}
\examples{
data("qPCR_data1")
result <- gam_analysis(qPCR_data1, smoothed = TRUE, fluo = 5, selections = c(
    "Well.Position", "Temperature", "Fluorescence", "Normalized"
))
read_tsar(result, code = 0)
output_data <- read_tsar(result, code = 2)

}
\seealso{
Other read_write_analysis: 
\code{\link{join_well_info}()},
\code{\link{write_tsar}()}
}
\concept{read_write_analysis}
