% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{remove_raw}
\alias{remove_raw}
\title{remove_raw}
\usage{
remove_raw(raw_data, removerange = NULL, removelist = NULL)
}
\arguments{
\item{raw_data}{dataframe; to be processed data}

\item{removerange}{list; list of 4 string, specifying range of wells to be removed}

\item{removelist}{list; list of indiviudal well numbers that need to be removed}
}
\value{
dataframe; data frame with specified well removed
}
\description{
removes selected curves with specified wells and range
}
