% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\title{screen function allows user to screen data for selection}
\usage{
screen(raw_data, checkrange = NULL, checklist = NULL)
}
\arguments{
\item{raw_data}{input raw_data}

\item{checkrange}{list type input identifying specific selections of well.
For example, if screening for only 6 wells of row A is needed, one can
specify the row letters and column numbers like this:
`checkrange = c("A", "C", "1", "8")`}

\item{checklist}{use this parameter to view selected Wells with full
Well names}
}
\value{
returns a ggplot graph colors by well IDs
}
\description{
screens multiple wells of data and prepares to assist identification of
  corrupted wells and odd out behaviors
}
\seealso{
Other data_preprocess: 
\code{\link{normalize}()}
}
\concept{data_preprocess}
