% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsa_analysis.R
<<<<<<< HEAD
\name{Tm_est}
\alias{Tm_est}
\title{find inflection point function}
\usage{
Tm_est(norm_data, min, max)
=======
\name{tm_est}
\alias{tm_est}
\title{find inflection point function}
\usage{
tm_est(norm_data, min, max)
>>>>>>> main
}
\arguments{
\item{norm_data}{data frame; data frame input containing derivative values
can only be data frames for one well; finding inflections points across
multiple wells require iteration through individual wells}

\item{min}{restricts finding to be above the given minimum temperature}

\item{max}{restricts finding to be below the given maximum temperature
parameter min and max can be used to remove messy or undesired data
for better accuracy in tm estimation; removing data is  before fitting the
model is more recommended than removing here}
}
\value{
integer; tm estimation
<<<<<<< HEAD

#examples
#well <- fitted %>%
#   filter(Well.Position == "A1")
#tm_est(A1)
#tm_est(A1, min = 30, max= 35)
=======
>>>>>>> main
}
\description{
looks for Tm temperature values by finding the inflection point in the
  normalized fluorescence data. The inflection point is approximated by
  locating the maximum first derivative
}
<<<<<<< HEAD
=======
\concept{tsa_analysis

#examples
#well <- fitted %>%
#   filter(Well.Position == "A1")
#tm_est(A1)
#tm_est(A1, min = 30, max= 35)}
>>>>>>> main
