% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsar.R
\name{tsar}
\alias{tsar}
\title{scalable time-series computations on 3D raster stacks}
\usage{
tsar(raster.name, workers, cores, out.name, out.bandnames = NULL,
  out.dtype = "FLT4S", separate = T, na.in = NA, na.out = -99,
  overwrite = F, verbose = T, nodelist = NULL, bandorder = "BSQ",
  maxmemory = 100, compress_tif = F, mask = NULL)
}
\arguments{
\item{raster.name}{a 3D raster object with dimensions in order lines-samples-time}

\item{workers}{a named list containing the functions for computation}

\item{cores}{the number of parallel processes per node}

\item{out.name}{the name of the output. Either a single file or a folder of separate files 
(determined by parameter \code{separate})}

\item{out.bandnames}{(optional) the names of the output bands; names are determined from the function names 
in \code{workers} if left empty}

\item{out.dtype}{the datatype of the written files.
See \code{\link[raster]{dataType}} for options.}

\item{separate}{should the resulting bands be written to individual files? Otherwise a single ENVI block is written.}

\item{na.in}{the pixel value for NA in \code{raster.name}}

\item{na.out}{the pixel value for NA in the output files}

\item{overwrite}{should the output files be overwritten if they already exist? If \code{separate} all output files are checked}

\item{verbose}{write detailed information on the progress of function execution?}

\item{nodelist}{the names of additional server computing nodes accessible via SSH without password}

\item{bandorder}{the output file pixel arrangement,one of 'BIL', 'BIP' or 'BSQ'}

\item{maxmemory}{the maximum memory in Mb used per node}

\item{compress_tif}{should the written GeoTiff files be compressed?}

\item{mask}{an additional file or raster layer; computations on raster.name are only performed where mask is 1, 
otherwise NA is returned for all resulting layers}
}
\value{
None
}
\description{
scalable time-series computations on 3D raster stacks
}
\seealso{
\code{\link[raster]{stack}}, \code{\link[raster]{calc}}, \code{\link[raster]{clusterR}}, \code{\link[snow]{makeCluster}}
}
