% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{view_deriv}
\alias{view_deriv}
\title{View Derivative Curves}
\usage{
view_deriv(norm_data)
}
\arguments{
\item{norm_data}{dataset input, analyzed must norm_deriv as a variable;
dataset qualifying norm_data or tsar_data both fulfills this parameter}
}
\value{
plotly object of derivative curves
}
\description{
Function reviews data by well and output a graph of the all derivatives
  wanted. Function called within graph_tsar function but also runnable
  outside.
}
\examples{
data("example_tsar_data")
view_deriv(example_tsar_data)

}
\seealso{
Other TSA Plots: 
\code{\link{TSA_boxplot}()},
\code{\link{TSA_compare_plot}()},
\code{\link{TSA_wells_plot}()},
\code{\link{animante_deriv}()},
\code{\link{get_legend}()},
\code{\link{graph_tsar}()}
}
\concept{TSA Plots}
