% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weed_raw.R
\name{weed_raw}
\alias{weed_raw}
\title{Weed Raw}
\usage{
weed_raw(raw_data, checkrange = NULL, checklist = NULL)
}
\arguments{
\item{raw_data}{The raw data for screening.}

\item{tobechecked}{The curves to be checked during screening.

#examples weed_raw(raw_data, tobechecked = c("A11", "A12"))}
}
\description{
The weed_raw function allows users to interact with a screening graph and select
curves to weed out before entering analysis.
}
