% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_norm.R
\name{analyze_norm}
\alias{analyze_norm}
\title{Analyze to Normalize}
\usage{
analyze_norm(raw_data)
}
\arguments{
\item{raw_data}{The raw data for analysis.}
}
\value{
shiny application
}
\description{
The analyze_norm function allows users to process analysis through an UI
    interface. Function wraps together all functions with in TSA_analysis
    family and read_write_analysis family.
}
\examples{
if (interactive()) {
    data("qPCR_data1")
    shiny::runApp(analyze_norm(qPCR_data1))
}

}
\seealso{
\code{\link[TSAR]{gam_analysis}}, \code{\link{read_tsar}},
  \code{\link{write_tsar}}, \code{\link{join_well_info}}
}
