% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_tsar.R
\name{graph_tsar}
\alias{graph_tsar}
\title{Graph tsar_data}
\usage{
graph_tsar(tsar_data = data.frame())
}
\arguments{
\item{tsar_data}{tsar data outputted by merge_norm or merge_tsa.
Parameter is optional. If no data is passed, access the merge panel
to merge norm_data into tsar_data.}
}
\value{
prompts separate app window for user interaction,
    does not return specific value; generates boxplot and compare plots
    according to user input
}
\description{
Function allows users to graph out tsar_data, building boxplot, compare
    plots, and curves by condition. Input of data as parameter is optional.
    graph_tsar wraps together all graphing functions and relative helper
    functions.
}
\examples{
if (interactive()) {
    data("example_tsar_data")
    shiny::runApp(graph_tsar(example_tsar_data))
}

}
\seealso{
\code{\link{TSA_boxplot}}, \code{\link{TSA_compare_plot}},
    \code{\link{condition_IDs}}, \code{\link{well_IDs}},
    \code{\link{merge_norm}}, \code{\link{TSA_Tms}},
    \code{\link{Tm_difference}}

Other TSA Plots: 
\code{\link{TSA_boxplot}()},
\code{\link{TSA_compare_plot}()},
\code{\link{TSA_wells_plot}()},
\code{\link{get_legend}()},
\code{\link{view_deriv}()}
}
\concept{TSA Plots}
