% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_well_info.R
\name{join_well_info}
\alias{join_well_info}
\title{well information input function}
\usage{
join_well_info(file_path, analysis_file, skips = 2, nrows = 95, type)
}
\arguments{
\item{file_path}{string; file path to read in the file}

\item{analysis_file}{data frame; data frame containing smoothed fluorescence
data and tm values}

\item{skips}{integer; number indicating the number of headers present in
input file, default set to 2 when file input is by_well
If the input follows the excel template, this parameter does not apply,
can be ignored}

\item{nrows}{integer; number indicating the number of rows the data is.
Default set to 95 assuming analysis on 96 well plate. If inputting by excel
template, this parameter does not apply, may be ignored.}

\item{type}{string; variable specifies the type of input read in.
type = "by_well" requires input of csv or txt files of three variables:
Well, Protein, Ligand.
type = "by_template" requires input of excel file following the template
format provided}
}
\value{
outputs data fram joining data information with well information

examples
#Analysis <- join.well_info("/Users/candygao/Desktop/qpcrresult/
                           CA_IP_HCB_2_20220110_134917_AnalysisResults.txt",
                           output_data, type = "by_well")
#Analysis <- join.well_info("~/Desktop/qpcrresult/Well Information Template.xlsx",
                         output_data,
                         type = "by_template")
}
\description{
reads in the ligand and protein information and joins them accordingly
to the big data frame for graphing purposes
}
