% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_to_tsar.R
\name{merge_norm}
\alias{merge_norm}
\title{Merge and format data for sequential applications of graphing functions}
\usage{
merge_norm(data, name, date)
}
\arguments{
\item{data}{A character vector specifying the file paths of the data files
or data frame objects of analysis data set.}

\item{name}{A character vector specifying the experiment names.}

\item{date}{A character vector specifying the dates.}
}
\description{
This function merges data of experiment replicates across different dates.
  It merges and produces information variables used to group wells of same
  set up.
}
\details{
This function merges and normalizes test data from multiple files.
  The lengths of the \code{data}, \code{name}, and \code{date} vectors
  must match, otherwise an error is thrown.
}
\seealso{
Other TSAR Formatting: 
\code{\link{TSA_Tms}()},
\code{\link{Tm_difference}()},
\code{\link{merge_tsa}()},
\code{\link{normalize_fluorescence}()},
\code{\link{rescale}()},
\code{\link{tsa_average}()}
}
\concept{TSAR Formatting}
