% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_to_tsar.R
\name{merge_norm}
\alias{merge_norm}
\title{Merge and format data for sequential applications of graphing functions}
\usage{
merge_norm(data, name, date)
}
\arguments{
\item{data}{A character vector specifying the file paths of the data files
or data frame objects of analysis data set.}

\item{name}{A character vector specifying the experiment names.}

\item{date}{A character vector specifying the dates.}
}
\value{
data frame in the format of tsat_data
}
\description{
This function merges data of experiment replicates across different dates.
  It merges and produces information variables used to group wells of same
  set up.
}
\details{
This function merges and normalizes test data from multiple files.
  The lengths of the \code{data}, \code{name}, and \code{date} vectors
  must match, otherwise an error is thrown.
}
\examples{
data("qPCR_data1")
result <- gam_analysis(qPCR_data1, smoothed = TRUE, fluo = 5)
data("Well_Information")
norm_data <- join_well_info( file_path = NULL, file = Well_Information,
  read_tsar(result, code = 2), type = "by_template")
norm_data <- na.omit(norm_data)
data("qPCR_data2")
result2 <- gam_analysis(qPCR_data1, smoothed = TRUE, fluo = 5)
norm_data2 <- join_well_info( file_path = NULL, file = Well_Information,
  read_tsar(result2, code = 2), type = "by_template")
norm_data2 <- na.omit(norm_data2)
tsar_data <- merge_norm(
  data = list(norm_data, norm_data2),
  name = c( "Thermal Shift_162.eds.csv", "Thermal Shift_168.eds.csv"),
  date = c("20230203", "20230209"))


}
\seealso{
Other TSAR Formatting: 
\code{\link{TSA_Tms}()},
\code{\link{Tm_difference}()},
\code{\link{merge_tsa}()},
\code{\link{normalize_fluorescence}()},
\code{\link{rescale}()},
\code{\link{tsa_average}()}
}
\concept{TSAR Formatting}
