% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{model_fit}
\alias{model_fit}
\title{Refit and calculate derivative function}
\usage{
model_fit(norm_data, model, smoothed)
}
\arguments{
\item{norm_data}{data frame; the raw data set input}

\item{model}{fitted model containing fitted values}

\item{smoothed}{inform whether data already contains smoothed model; set
\code{smoothed = "Flourescence"} to calculate norm_deriv upon the
existing model}
}
\value{
data frame; with calculated derivative columns
}
\description{
Model_fit calculates derivatives by refitting model onto data
}
\concept{data_preprocess

#examples
#fitted <- model_fit(data = gammodel, model = fitted)}
