% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{model_fit}
\alias{model_fit}
\title{Refit and calculate derivative function}
\usage{
model_fit(norm_data, model)
}
\arguments{
\item{norm_data}{data frame; the raw data set input}

\item{model}{fitted model containing fitted values}
}
\value{
data frame; with calculated derivative columns

#examples
#fitted <- modelfit(data = gammodel, model = fitted)
}
\description{
Model_fit calculates derivatives by refitting model onto data
}
