% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{model_gam}
\alias{model_gam}
\title{modeling function}
\usage{
model_gam(norm_data, x, y)
}
\arguments{
\item{norm_data}{data frame input}

\item{x}{temperature column}

\item{y}{normalized fluorescence column}
}
\value{
gam model fitted with formula set to y ~ s(x, bs = "cs")
<<<<<<< HEAD
#example
#gammodel <- model_gam(Data = normalized,
                         x = normalized$temperature,
                         y = normalized$fluorescence)
=======
>>>>>>> main
}
\description{
function finds fitted fluorescence values
}
<<<<<<< HEAD
=======
\concept{data_preprocess
#example
#gammodel <- model_gam(Data = normalized,
                         x = normalized$temperature,
                         y = normalized$fluorescence)}
>>>>>>> main
