% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{model_gam}
\alias{model_gam}
\title{modeling function}
\usage{
model_gam(norm_data, x, y)
}
\arguments{
\item{norm_data}{data frame input}

\item{x}{temperature column}

\item{y}{normalized fluorescence column}
}
\value{
gam model fitted with formula set to y ~ s(x, bs = "ad")
}
\description{
function finds fitted fluorescence values
}
\examples{
data("qPCR_data1")
test <- subset(qPCR_data1, Well.Position == "A01")
test <- normalize(test, fluo = 5, selected = c(
    "Well.Position", "Temperature",
    "Fluorescence", "Normalized"
))
model_gam(test, x = test$Temperature, y = test$Normalized)

}
\seealso{
Other data_preprocess: 
\code{\link{model_fit}()},
\code{\link{normalize}()},
\code{\link{remove_raw}()},
\code{\link{screen}()},
\code{\link{view_model}()},
\code{\link{weed_raw}()}
}
\concept{data_preprocess}
