% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_analysis.R
\name{read_tsar}
\alias{read_tsar}
\title{read result}
\usage{
read_tsar(result, code)
}
\arguments{
\item{result}{list; input uses resulting output of gam_analysis() function}

\item{code}{integer; code = 0 returns only the tm value by wells
code = 1 returns the data table with the fitted values
code = 2 returns the combination of 0 and 1}
}
\value{
output files with select dataset
}
\description{
reads previous pipeline output lists and organizes them into separate
  data frames
}
\examples{
#read_tsar(result, code = 0)
#output_data <- read_tsar(result, code = 2)

}
\concept{read_write_analysis}
