% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsa_curves.R
\name{tsa_compare_plot}
\alias{tsa_compare_plot}
\title{Compare TSA curves to control}
\usage{
tsa_compare_plot(
  tsa_data,
  control_condition,
  y = "Fluorescence",
  show_Tm = FALSE,
  title_by = "both",
  digits = 1
)
}
\arguments{
\item{tsa_data}{a data frame that is merged and generated by
<<<<<<< HEAD
TSAR::merge_TSA(). If \code{y = "RFU"}, tsa_data must also be
=======
TSAR::merge_tsa(). If \code{y = "RFU"}, tsa_data must also be
>>>>>>> main
generated by TSAR::normalize_fluorescence. The Temperature column will be
rounded and the average & sd of each rounded temperature is calculated.}

\item{control_condition}{character string matching a Condition ID. Must be
equal to a value within tsa_data$condition_ID. See unique condition IDs
with \code{\link{condition_IDs}}.}

\item{y}{character string; c("Fluorescence", "RFU").
When \code{y = "Fluorescence"}, the original Fluorescence data from
TSAR::read_raw_data() is averaged. When \code{y = "RFU"}, the average is
calculated by the rescaled fluorescence.}

\item{show_Tm}{logical; \code{show_Tm = FALSE} by default. When TRUE, the Tm
is displayed on the plot. When FALSE, the Tm is not added to the plot.}

\item{title_by}{character string; c("ligand", "protein", "both").
Automatically names the plots by the specified condition category.}

<<<<<<< HEAD
\item{digits}{integer; the number of decimal places to round for change in Tm
=======
\item{digits}{integer; the numer of decimal places to round for change in Tm
>>>>>>> main
calculations displayed in the subtitle of each plot..}
}
\value{
Generates a number of ggplot objects equal to the number of unique
 Condition IDs present in the input data.
}
\description{
Generate a number of plots based on the input data to compare the average
and standard deviation (sd) of each unique condition to a specified
control condition. To see all conditions use \code{condition_IDs(tsa_data)}.
}
\seealso{
<<<<<<< HEAD
\code{\link{merge_TSA}} and \code{\link{normalize_fluorescence}}
=======
\code{\link{merge_tsa}} and \code{\link{normalize_fluorescence}}
>>>>>>> main
  for preparing data. See \code{\link{tsa_average}} and
  \code{\link{get_legend}} for details on function parameters.
  See \code{\link{TSA_wells_plot}} for individual curves of the averaged
  conditions shown.

Other TSA Plots: 
\code{\link{TSA_boxplot}()},
<<<<<<< HEAD
\code{\link{TSA_wells_plot}()}
=======
\code{\link{TSA_wells_plot}()},
\code{\link{get_legend}()}
>>>>>>> main
}
\concept{TSA Plots}
