% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsar.R
\name{tsar}
\alias{tsar}
\title{\code{tsar} scalable time-series computations on 3D raster stacks}
\usage{
tsar(raster.name, workers, cores, out.name, out.bandnames = NULL,
  out.dtype = "FLT4S", separate = T, na.in = NA, na.out = -99,
  overwrite = T, verbose = T, nodelist = NULL)
}
\arguments{
\item{raster.name}{a 3D raster object with dimensions in order lines-samples-time}

\item{workers}{a named list containing the functions for computation}

\item{cores}{the number of parallel processes}

\item{out.name}{the name of the output. Either a single file or a folder separate files (determined by parameter separate)}

\item{out.bandnames}{(optional) the names of the output bands; names are determined from the function names in workers if left empty}

\item{out.dtype}{the datatype of the written files}

\item{separate}{should the resulting band be written to indivudual files? Otherwise a single ENVI block is written.}

\item{na.in}{the pixel value for NA in raster.name}

\item{na.out}{the pixel value for NA in the output files}

\item{overwrite}{should the outputfiles be overwritten if they already exist? If separate all output files are checked}

\item{verbose}{write detailed information on the progress of function execution}

\item{nodelist}{the names of additional server computing nodes accessible via SSH without password}
}
\value{
None
}
\description{
\code{tsar} scalable time-series computations on 3D raster stacks
}
\seealso{
\code{\link{raster}}, \code{\link{foreach}}, \code{\link{doSNOW}}
}
