% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{view_model}
\alias{view_model}
\title{View Model}
\usage{
view_model(norm_data)
}
\arguments{
\item{norm_data}{dataset input, data should match the needs of norm_data}
}
\value{
ggplot
}
\description{
View Model
}
